update rpc health when provider is requested
